+++ src/dml/delete/mssql_delete_target.cpp
@@ -1,10 +1,11 @@
 #include "dml/delete/mssql_delete_target.hpp"
 #include "duckdb/common/string_util.hpp"
+#include "query/mssql_sql_params.hpp"
 
 namespace duckdb {
 
 string MSSQLDeleteTarget::GetFullyQualifiedName() const {
-	return "[" + schema_name + "].[" + table_name + "]";
+	return mssql::QuoteIdentifier(schema_name) + "." + mssql::QuoteIdentifier(table_name);
 }
 
 idx_t MSSQLDeleteTarget::GetParamsPerRow() const {